# a9_openapi.yaml
openapi: 3.0.0
info:
  version: "2.0"
  title: "LBAW SLCN Web API"
  description: "Web Resources Specification (A9) for SLCN"
servers:
  - url: http://lbaw2286.lbaw.fe.up.pt
    description: Production server
...
paths:
  /login:
    get:
      operationId: R101
      summary: "R101: LoginForm"
      description: "Provide login form. Access: PUB"
      tags:
        - "M01: Authentication"
      responses:
        "200":
          description: "Ok. Show Log-in UI10"
    post:
      operationId: R102
      summary: "R102: Login Action"
      description: "Processes the login form submission. Access: PUB"
      tags:
        - "M01: Authentication"

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email: # <!--- form field name
                  type: string
                password: # <!--- form field name
                  type: string
              required:
                - email
                - password

      responses:
        "302":
          description: "Redirect after processing the login credentials."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful authentication. Redirect main page."
                  value: "/"
                302Error:
                  description: "Failed authentication. Redirect to login form."
                  value: "/login"
...
  /register:
    get:
      operationId: R104
      summary: "R104: Register Form"
      description: "Provide new user registration form. Access: PUB"
      tags:
        - "M01: Authentication"
      responses:
        "200":
          description: "Ok. Show Sign-Up UI09"

    post:
      operationId: R105
      summary: "R105: Register Action"
      description: "Processes the new user registration form submission. Access: PUB"
      tags:
        - "M01: Authentication"

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        "302":
          description: "Redirect after processing the new user information."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful authentication. Redirect to main page."
                  value: "/"
                302Failure:
                  description: "Failed authentication. Redirect to login form."
                  value: "/login"
  /profile/{id}:
    get:
      operationId: R201
      summary: "R201: Gets a profile page of a user."
      description: "Displays profile page of a user. Access: USR"
      tags:
        - "M02: Users"
      responses:
        "200":
          description: "Ok. Show Profile page [UI03]"
        "302":
          description: "Redirect."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "Redirect to login page."
                  value: "/login"
  /edit_profile/{id}:
    post:
      operationId: R202
      summary: "R202: Displays User edit page."
      description: "Display edit page for a user's profile. Access: AUTH, ADM"
      tags:
        - "M02: Users"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "302":
          description: "Redirect."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Redirect to profile edit page."
                  value: "/edit_profile/{id}"
                302Failure:
                  description: "Redirect to profile page."
                  value: "/profile/{id}"
  /api/edit_profile/{id}:
    post:
      operationId: R203
      summary: "R203: Edits a user's profile."
      description: "Processes and edits a user's profile. Access: AUTH, ADM"
      tags:
        - "M02: Users"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                username:
                  type: string
                country:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - id
                - username
                - country
                - email
                - password
      responses:
        "302":
          description: "Redirect."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Redirect to profile page."
                  value: "/profile/{id}"
                302Failure:
                  description: "Redirect to profile edit page."
                  value: "/edit_profile/{id}"
  /change_admin/{id}:
    post:
      operationId: R204
      summary: "R204: Changes a users admin role."
      description: "Changes a users admin role. If the user is an admin, they lose the role and vice-versa. Access: ADM"
      tags:
        - "M02: Users"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "302":
          description: "Redirect."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Redirect to profile page."
                  value: "/profile/{id}"
                302Failure:
                  description: "Redirect to profile page."
                  value: "/profile/{id}"
...
  /news/{id}:
    get:
      operationId: R302
      summary: "R302: Shows news item."
      description: "Obtains page with the selected news item. Access; USR"
      tags:
        - "M03: News"
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: "Ok. Show news item [UI02]"
        "302":
          description: "User not logged in. Redirect to login form"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to login page."
                  value: "/login"
  /api/news:
    post:
      operationId: R303
      summary: "R303: News creation action."
      description: "Processes the new news item form submition and creates the news item. Access; USR"
      tags:
        - "M03: News"
      requestBody:
        required: true
        content:
          application/x-www-form-urlenconded:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                picture:
                  type: object
                id_author:
                  type: integer
              required:
                - title
                - content
                - id_author

      responses:
        "302":
          description: "Redirect after processing new news item."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful news creation. Redirection to news item."
                  value: "/news/{id}"
                302Failure:
                  description: "Failure in news creation. Redirection to home page."
                  value: "/"
  /api/news/{news_id}:
    post:
      operationId: R304
      summary: "R304: Deletes a news item"
      description: "Deletes a news item; Access: AUTH, ADM"
      tags:
        - "M03: News"
      requestBody:
        required: true
        content:
          application/x-www-form-urlenconded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "302":
          description: "Redirect after processing and deleting news item."
          headers:
            location:
              schema:
                type: string
              examples:
                302Sucess:
                  description: "Successful news deletion. Redirection to home page."
                  value: "/"
                302Failure:
                  description: "Failure in news deletion. Redirection to news page."
                  value: "/news/{id}"
  /api/news/update/:
    post:
      operationId: R305
      summary: "R305: Edits a news item"
      description: "Processes and edits a news item; Access: AUTH, ADM"
      tags:
        - "M03: News"
      requestBody:
        required: true
        content:
          application/x-www-form-urlenconded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                picture:
                  type: object
                id_author:
                  type: integer
              required:
                - id
                - title
                - content
                - id_author
      responses:
        "302":
          description: "Redirect after editing news item."
          headers:
            location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful news update. Redirection to news page."
                  value: "/news/{id}"
                302Failure:
                  description: "Failure in news update. Redirection to news page."
                  value: "/news/{id}"
  /search:
    get:
      operationId: R701
      summary: "R701: Search for news posts or users"
      description: "Search for news posts or users. Access: PUB"
      tags:
        - "M07: Search"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                query:
                  type: string
              required:
                - query
      responses:
        "302":
          description: "Redirect after processing the query."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful search. Redirect to search results page.[UI05]"
                  value: "/search/{query}"
                302Failure:
                  description: "Failed search. Redirect to home page."
                  value: "/"
        "400":
          description: "Bad Request"
  /following:
    get:
      operationId: R306
      summary: "R306: Show user curated news feed"
      description: "Search for news posts or users. Access: AUTH, ADM"
      tags:
        - "M03: News"
      responses:
        "200":
          description: "Ok, show curated news feed"
        "302":
          description: "User not logged in. Redirect to main feed"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to main feed."
                  value: "/"
...
  /rte:
    get:
      operationId: R307
      summary: "R307: Show rich text news creation page"
      description: "Shows rich text news creation page. Access: AUTH, ADM"
      tags:
        - "M03: News"
      responses:
        "200":
          description: "Ok, show rich text news create page"
        "302":
          description: "User not logged in. Redirect to login page"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to login page."
                  value: "/login"
  /rte/{id}:
    get:
      operationId: R308
      summary: "R308: Show rich text news edit page"
      description: "Shows rich text news edit page. Access: AUTH, ADM"
      tags:
        - "M03: News"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, show rich text news edit page"
        "302":
          description: "User not logged in. Redirect to login page"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to login page."
                  value: "/login"
  /api/delete_profile/{id}:
    post:
      operationId: R202
      summary: "R202: Deletes user account"
      description: "Deletes user account. Access: AUTH, ADM"
      tags:
        - "M02: Users"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, deletes user account"
        "401":
          description: "Anauthorized access"
          headers:
            Location:
              schema:
                type: string
  /api/follow:
    post:
      operationId: R203
      summary: "R202: Follows a user"
      description: "Follows a user. Access: AUTH, ADM"
      tags:
        - "M02: Users"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id1:
                  type: integer
                id2:
                  type: integer
              required:
                - id1
                - id2
      responses:
        "200":
          description: "Ok, Follows user"
        "401":
          description: "Anauthorized access"
          headers:
            Location:
              schema:
                type: string
  /api/unfollow:
    post:
      operationId: R204
      summary: "R204: unFollows a user"
      description: "unFollows a user. Access: AUTH, ADM"
      tags:
        - "M02: Users"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id1:
                  type: integer
                id2:
                  type: integer
              required:
                - id1
                - id2
      responses:
        "200":
          description: "Ok, follows user"
        "401":
          description: "Anauthorized access"
          headers:
            Location:
              schema:
                type: string
  /follow_list/{id}:
    get:
      operationId: R205
      summary: "R205: Show followed users list"
      description: "Show followed users list. Access: AUTH, ADM"
      tags:
        - "M02: Users"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, show followed users list"
        "302":
          description: "User not logged in. Redirect to login page"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to login page."
                  value: "/login"
  /api/new_comment:
    post:
      operationId: R401
      summary: "R401: Creates a comment"
      description: "Creates a comment. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id_news:
                  type: integer
                content:
                  type: string
              required:
                - id_news
                - content
      responses:
        "200":
          description: "Ok, creates comment. Refreshes news page."
        "302":
          description: "User not logged in. Redirect to login page"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to login page."
                  value: "/login"
  /api/del_comment:
    post:
      operationId: R402
      summary: "R402: Deletes a comment"
      description: "Deletes a comment. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, deletes a comment. Refreshes news page."
        "302":
          description: "User not logged in. Redirect to login page"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to login page."
                  value: "/login"
  /api/edit_comment:
    post:
      operationId: R403
      summary: "R403: Edits a comment"
      description: "Edits a comment. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                content:
                  type: string
              required:
                - id
                - content
      responses:
        "200":
          description: "Ok, edits comment. Refreshes news page."
        "302":
          description: "User not logged in. Redirect to login page"
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: "User not logged in. Redirect to login page."
                  value: "/login"
  /api/getReplies:
    post:
      operationId: R404
      summary: "R404: Retrieves replies to a comment."
      description: "Retrieves replies to a comment with AJAX. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                id_comment:
                  type: integer
              required:
                - id
                - id_comment
      responses:
        "200":
          description: "Ok, returns replies as JSON."
        "401":
          description: "User not logged in. Access Denied"
  /api/createReply:
    post:
      operationId: R405
      summary: "R405: Creates a reply to a comment."
      description: "Creates a reply to a comment with AJAX. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                news_id:
                  type: integer
                id_comment:
                  type: integer
              required:
                - news_id
                - id_comment
      responses:
        "200":
          description: "Ok, Reply created."
        "401":
          description: "User not logged in. Access Denied"
  /api/editReply:
    post:
      operationId: R406
      summary: "R406: Edits a reply to a comment."
      description: "Edits a reply to a comment with AJAX. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                content:
                  type: string
              required:
                - id
                - content
      responses:
        "200":
          description: "Ok, Reply Edited Successfuly."
        "401":
          description: "User not logged in. Access Denied"
  /api/delReply:
    delete:
      operationId: R407
      summary: "R407: Deletes a reply to a comment."
      description: "Deletes a reply to a comment with AJAX. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, Reply Deleted Successfuly."
        "401":
          description: "User not logged in. Access Denied"
  /api/vote/newsCreate:
    post:
      operationId: R309
      summary: "R309: Creates a vote on a news."
      description: "Creates a vote on a news with AJAX. Access: AUTH, ADM"
      tags:
        - "M03: News"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                is_liked:
                  type: integer
              required:
                - id
                - is_liked
      responses:
        "200":
          description: "Ok, Vote Created."
        "401":
          description: "User not logged in. Access Denied"
  /api/vote/newsUpdate:
    post:
      operationId: R310
      summary: "R310: Updates a vote on a news."
      description: "Updates a vote on a news with AJAX. Access: AUTH, ADM"
      tags:
        - "M03: News"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                is_liked:
                  type: integer
              required:
                - id
                - is_liked
      responses:
        "200":
          description: "Ok, Vote Updated."
        "401":
          description: "User not logged in. Access Denied"
  /api/vote/newsDelete:
    delete:
      operationId: R311
      summary: "R311: Deletes a vote on a news."
      description: "Deletes a vote on a news with AJAX. Access: AUTH, ADM"
      tags:
        - "M03: News"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, Vote Deleted."
        "401":
          description: "User not logged in. Access Denied"
  /api/vote/commentCreate:
    post:
      operationId: R408
      summary: "R408: Creates a vote on a comment."
      description: "Creates a vote on a comment with AJAX. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                is_liked:
                  type: integer
              required:
                - id
                - is_liked
      responses:
        "200":
          description: "Ok, Vote Created."
        "401":
          description: "User not logged in. Access Denied"
  /api/vote/commentUpdate:
    post:
      operationId: R409
      summary: "R409: Updates a vote on a comment."
      description: "Updates a vote on a comment with AJAX. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                is_liked:
                  type: integer
              required:
                - id
                - is_liked
      responses:
        "200":
          description: "Ok, Vote Updated."
        "401":
          description: "User not logged in. Access Denied"
  /api/vote/commentDelete:
    delete:
      operationId: R410
      summary: "R410: Deletes a vote on a comment."
      description: "Deletes a vote on a comment with AJAX. Access: AUTH, ADM"
      tags:
        - "M04: Comments"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, Vote Deleted."
        "401":
          description: "User not logged in. Access Denied"
  /api/tag/propose:
    post:
      operationId: R501
      summary: "R501: Creates a tag proposal."
      description: "Creates a tag proposal with AJAX. Access: AUTH, ADM"
      tags:
        - "M05: Tags"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        "200":
          description: "Ok, Tag Proposal Created."
        "401":
          description: "User not logged in. Access Denied"
  /api/tag/create:
    post:
      operationId: R502
      summary: "R502: Creates a tag and handles the proposal."
      description: "Creates a tag and handles the proposal with AJAX. Access: AUTH, ADM"
      tags:
        - "M05: Tags"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      responses:
        "200":
          description: "Ok, Tag Created, Proposal Handled."
        "401":
          description: "User not logged in. Access Denied"
...
  /api/sendnotifications:
    post:
      operationId: R601
      summary: "R601: Send a notification"
      description: "Sends a notification using pusher. Access: AUTH, PUB"
      tags:
        - "M06: Notifications"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                type:
                  type: string
                user_id:
                  type: integer
                receiver_id:
                  type: integer
              required:
                - id
                - type
                - user_id
                - receiver_id
      responses:
        "200":
          description: "Ok, Notification Sent."
        "401":
          description: "User not logged in. Access Denied"
...